# tia/dua-tia-configuration-manager/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/tia/api/v1/document-types": {
            "get": {
                "tags": [
                    "document-type-controller"
                ],
                "operationId": "getDocumentTypes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentType"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "document-type-controller"
                ],
                "operationId": "saveDocumentType",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentType"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentType"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/document-repositories": {
            "get": {
                "tags": [
                    "document-repository-controller"
                ],
                "operationId": "getDocumentRepositoryList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentRepository"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "document-repository-controller"
                ],
                "operationId": "saveDocumentRepository",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRepository"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentRepository"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/credentials": {
            "get": {
                "tags": [
                    "credential-controller"
                ],
                "operationId": "getCredentialList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Credential"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "credential-controller"
                ],
                "operationId": "saveCredential",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Credential"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Credential"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/bulk-configuration": {
            "get": {
                "tags": [
                    "bulk-configuration-controller"
                ],
                "operationId": "getMethodName",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkConfiguration"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "bulk-configuration-controller"
                ],
                "operationId": "loadConfiguration",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkConfiguration"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "bulk-configuration-controller"
                ],
                "operationId": "dropConfiguration",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/tia/api/v1/document-types/{id}": {
            "get": {
                "tags": [
                    "document-type-controller"
                ],
                "operationId": "getDocumentType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentType"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "document-type-controller"
                ],
                "operationId": "deleteDocumentType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/document-repositories/{id}": {
            "get": {
                "tags": [
                    "document-repository-controller"
                ],
                "operationId": "getDocumentRepository",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentRepository"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "document-repository-controller"
                ],
                "operationId": "deleteDocumentRepository",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/credentials/{id}": {
            "get": {
                "tags": [
                    "credential-controller"
                ],
                "operationId": "getDocumentRepository_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Credential"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "credential-controller"
                ],
                "operationId": "deleteCredential",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RepositorySearchBinding": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "initiativeId": {
                        "type": "string"
                    },
                    "documentTypeId": {
                        "type": "string"
                    },
                    "keywordDictionaryId": {
                        "type": "string"
                    },
                    "documentRepositoryIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "Comments": {
                "type": "object",
                "properties": {
                    "generatedBy": {
                        "type": "string"
                    },
                    "generationTime": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "KeywordDictionary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "shortName": {
                        "type": "string"
                    },
                    "get_comments": {
                        "$ref": "#/components/schemas/Comments"
                    },
                    "dictionaryDefinition": {
                        "$ref": "#/components/schemas/KeywordDictionaryDefinition"
                    }
                }
            },
            "KeywordDictionaryDefinition": {
                "type": "object",
                "properties": {
                    "keywords": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "skipKeywords": {
                        "type": "boolean"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/KeywordDictionaryDefinition"
                        }
                    }
                }
            },
            "Initiative": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "shortName": {
                        "type": "string"
                    }
                }
            },
            "DocumentType": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "shortName": {
                        "type": "string"
                    }
                }
            },
            "DocumentRepository": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "driver": {
                        "type": "string"
                    },
                    "credentials": {
                        "type": "string"
                    },
                    "frequency": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "pageLimit": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pauseBetweenPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "lastVisit": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Credential": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
            "BulkConfiguration": {
                "type": "object",
                "properties": {
                    "initiatives": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Initiative"
                        }
                    },
                    "documentTypes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DocumentType"
                        }
                    },
                    "documentRepositories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DocumentRepository"
                        }
                    },
                    "keywordDictionaries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/KeywordDictionary"
                        }
                    },
                    "repositorySearchBindingReferences": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RepositorySearchBindingReference"
                        }
                    },
                    "repositorySearchBindings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RepositorySearchBinding"
                        }
                    }
                }
            },
            "RepositorySearchBindingReference": {
                "type": "object",
                "properties": {
                    "initiativeIdx": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "documentTypeIdx": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "keywordDictionaryIdx": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "documentRepositoryIdxs": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                },
                "required": [
                    "documentRepositoryIdxs",
                    "documentTypeIdx",
                    "initiativeIdx",
                    "keywordDictionaryIdx"
                ]
            }
        }
    }
}